[fix]ディレクトリ構造を変更
